moving tests into test module
